+++ lib/src/core/better_player_controller.dart
@@ -42,6 +42,13 @@ class BetterPlayerController {
   ///between flutter high level code and lower level native code.
   VideoPlayerController? videoPlayerController;
 
+  ///Controls configuration
+  late BetterPlayerControlsConfiguration _betterPlayerControlsConfiguration;
+
+  ///Controls configuration
+  BetterPlayerControlsConfiguration get betterPlayerControlsConfiguration =>
+      _betterPlayerControlsConfiguration;
+
   ///Expose all active eventListeners
   List<Function(BetterPlayerEvent)?> get eventListeners =>
       _eventListeners.sublist(1);
@@ -139,6 +146,9 @@ class BetterPlayerController {
   ///in configuration.
   double? _overriddenAspectRatio;
 
+  ///Overridden fit which will be used instead of fit passed in configuration.
+  BoxFit? _overriddenFit;
+
   ///Was Picture in Picture opened.
   bool _wasInPipMode = false;
 
@@ -205,6 +215,8 @@ class BetterPlayerController {
     this.betterPlayerPlaylistConfiguration,
     BetterPlayerDataSource? betterPlayerDataSource,
   }) {
+    this._betterPlayerControlsConfiguration =
+        betterPlayerConfiguration.controlsConfiguration;
     _eventListeners.add(eventListener);
     if (betterPlayerDataSource != null) {
       setupDataSource(betterPlayerDataSource);
@@ -1027,6 +1039,19 @@ class BetterPlayerController {
     return _overriddenAspectRatio ?? betterPlayerConfiguration.aspectRatio;
   }
 
+  // ignore: use_setters_to_change_properties
+  ///Setup overridden fit.
+  void setOverriddenFit(BoxFit fit) {
+    _overriddenFit = fit;
+  }
+
+  ///Get fit used in current video. If fit is null, then fit from
+  ///BetterPlayerConfiguration will be used. Otherwise [_overriddenFit] will be
+  ///used.
+  BoxFit getFit() {
+    return _overriddenFit ?? betterPlayerConfiguration.fit;
+  }
+
   ///Enable Picture in Picture (PiP) mode. [betterPlayerGlobalKey] is required
   ///to open PiP mode in iOS. When device is not supported, PiP mode won't be
   ///open.
@@ -1237,6 +1262,13 @@ class BetterPlayerController {
       BetterPlayerDataSource betterPlayerDataSource) async {
     return VideoPlayerController?.stopPreCache(betterPlayerDataSource.url,
         betterPlayerDataSource.cacheConfiguration?.key);
+  }
+
+  /// Sets the new [betterPlayerControlsConfiguration] instance in the
+  /// controller.
+  void setBetterPlayerControlsConfiguration(
+      BetterPlayerControlsConfiguration betterPlayerControlsConfiguration) {
+    this._betterPlayerControlsConfiguration = betterPlayerControlsConfiguration;
   }
 
   /// Add controller internal event.