+++ lib/src/core/better_player_with_controls.dart
@@ -25,7 +25,7 @@ class _BetterPlayerWithControlsState extends State<BetterPlayerWithControls> {
       widget.controller!.betterPlayerConfiguration.subtitlesConfiguration;
 
   BetterPlayerControlsConfiguration get controlsConfiguration =>
-      widget.controller!.betterPlayerConfiguration.controlsConfiguration;
+      widget.controller!.betterPlayerControlsConfiguration;
 
   final StreamController<bool> playerVisibilityStreamController =
       StreamController();
@@ -134,7 +134,7 @@ class _BetterPlayerWithControlsState extends State<BetterPlayerWithControls> {
             angle: rotation * pi / 180,
             child: _BetterPlayerVideoFitWidget(
               betterPlayerController,
-              betterPlayerController.betterPlayerConfiguration.fit,
+              betterPlayerController.getFit(),
             ),
           ),
           betterPlayerController.betterPlayerConfiguration.overlay ??