+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.81
+* Fixed full screen button padding in material controls.
+* Added `setBetterPlayerControlsConfiguration` in `BetterPlayerController`.
+* Added `setOverriddenFit` in `BetterPlayerController`.
+
 ## 0.0.80
 * Removed pedantic dependency.
 * Updated dependencies.